+++ Cargo.toml
@@ -1,14 +1,16 @@
 [package]
 name = "locator"
-version = "1.0.0"
+version = "1.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+alphanumeric-sort = "1.5.3"
 getset = "0.1.2"
 lazy_static = "1.4.0"
+pretty_assertions = "1.4.0"
 regex = "1.6.0"
+schemars = "0.8.19"
 serde = { version = "1.0.140", features = ["derive"] }
 strum = { version = "0.24.1", features = ["derive"] }
 thiserror = "1.0.31"
